modify test of videos

+++ tests/testthat/test_load_save_videos.R
@@ -4,14 +4,17 @@ library(Rcpp)
 test_that("load_and_save_videos",{
     if (imager:::has.ffmpeg())
     {
-        im <- imfill(10,10,3,val=c(0,0,0))
+        if (!is.na(system("ffmpeg -verion", intern = TRUE)[1])) 
+        {
+            im <- imfill(10,10,3,val=c(0,0,0))
         
-        ff <- tempfile(fileext=".mp4")
-        save.video(im,ff)
-        expect(file_test("-f",ff),TRUE)
-        im2 <- load.video(ff)
-        unlink(ff)
-        expect_equal(im,im2)
+            ff <- tempfile(fileext=".mp4")
+            save.video(im,ff)
+            expect(file_test("-f",ff),TRUE)
+            im2 <- load.video(ff)
+            unlink(ff)
+            expect_equal(im,im2)
+        } 
     }
     else
     {
